oauth flow required rules
